MPDocBench/tools/idp_tools/chandra/pyproject.toml
[project]
name = "chandra-ocr"
version = "0.2.0"
description = "OCR model that converts documents to markdown, HTML, or JSON."
readme = "README.md"
requires-python = ">=3.10"
license = {text = "Apache-2.0"}
authors = [
    {name = "Datalab", email = "hi@datalab.to"}
]
keywords = ["ocr", "pdf", "markdown", "layout"]
dependencies = [
    "beautifulsoup4>=4.14.2",
    "click>=8.0.0",
    "filetype>=1.2.0",
    "markdownify==1.1.0",
    "openai>=2.2.0",
    "pillow>=10.2.0",
    "pydantic>=2.12.0",
    "pydantic-settings>=2.11.0",
    "pypdfium2>=4.30.0",
    "python-dotenv>=1.1.1",
    "six>=1.17.0",
]

[project.optional-dependencies]
hf = [
    "torch>=2.8.0",
    "torchvision>=0.23.0",
    "transformers>=5.2.0",
    "accelerate>=1.11.0",
]
app = [
    "streamlit>=1.50.0",
]
all = [
    "chandra-ocr[hf,app]",
]

[build-system]
requires = ["setuptools>=61"]  # or "setuptools>=61", "flit-core", etc.
build-backend = "setuptools.build_meta"

[project.scripts]
chandra = "chandra.scripts.cli:main"
chandra_app = "chandra.scripts.run_app:main"
chandra_screenshot = "chandra.scripts.screenshot_app:main"
chandra_vllm = "chandra.scripts.vllm:main"

[tool.setuptools.packages.find]
include = ["chandra*"]

[dependency-groups]
dev = [
    "chandra-ocr[all]",
    "flask>=3.0.0",
    "pre-commit>=4.3.0",
    "pytest>=8.4.2",
]
